COM-1839 - create home component
