+++ src/AppContainer/index.tsx
@@ -1,12 +1,10 @@
 import React, { useEffect, useContext, useRef } from 'react';
-import { StatusBar, Text, View } from 'react-native';
+import { StatusBar, View } from 'react-native';
 import { connect } from 'react-redux';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import * as Analytics from 'expo-firebase-analytics';
+import Analytics from '../core/helpers/analytics';
 import asyncStorage from '../core/helpers/asyncStorage';
-import ProfileDetails from '../screens/profile/Profile';
 import ProfileEdition from '../screens/profile/ProfileEdition';
 import Camera from '../screens/Camera';
 import ImagePickerManager from '../screens/ImagePickerManager';
@@ -15,10 +13,8 @@ import { navigationRef } from '../navigationRef';
 import Authentication from '../screens/Authentication';
 import EmailForm from '../screens/EmailForm';
 import CreateAccount from '../screens/CreateAccount';
-import Catalog from '../screens/explore/Catalog';
 import BlendedAbout from '../screens/explore/BlendedAbout';
 import ElearningAbout from '../screens/explore/ELearningAbout';
-import CourseList from '../screens/courses/CourseList';
 import CourseProfile from '../screens/courses/CourseProfile';
 import SubProgramProfile from '../screens/courses/SubProgramProfile';
 import CardContainer from '../screens/courses/CardContainer';
@@ -30,51 +26,7 @@ import { UserType } from '../types/UserType';
 import styles from './styles';
 import PasswordEdition from '../screens/profile/PasswordEdition';
 import PasswordReset from '../screens/PasswordReset';
-import CatalogIcon from '../../assets/icons/CatalogIcon';
-import CatalogSelectedIcon from '../../assets/icons/CatalogSelectedIcon';
-import CoursesIcon from '../../assets/icons/CoursesIcon';
-import CoursesSelectedIcon from '../../assets/icons/CoursesSelectedIcon';
-import ProfileIcon from '../../assets/icons/ProfileIcon';
-import ProfileSelectedIcon from '../../assets/icons/ProfileSelectedIcon';
-
-const Tab = createBottomTabNavigator();
-
-interface tabBarProps {
-  focused: boolean
-}
-
-const Home = () => {
-  const style = styles();
-
-  return (
-    <Tab.Navigator tabBarOptions={{ showLabel: false, style: style.tabBar }} initialRouteName="Courses">
-      <Tab.Screen name="Catalog" component={Catalog} options={{
-        tabBarIcon: ({ focused }: tabBarProps) => (focused
-          ? <View style={style.iconContainer}>
-            <CatalogSelectedIcon />
-            <Text style={style.iconText}>Explorer</Text>
-          </View>
-          : <CatalogIcon style={style.iconContainer} />),
-      }} />
-      <Tab.Screen name="Courses" component={CourseList} options={{
-        tabBarIcon: ({ focused }: tabBarProps) => (focused
-          ? <View style={style.iconContainer}>
-            <CoursesSelectedIcon />
-            <Text style={style.iconText}>Mes formations</Text>
-          </View>
-          : <CoursesIcon style={style.iconContainer} />),
-      }} />
-      <Tab.Screen name="Profile" component={ProfileDetails} options={{
-        tabBarIcon: ({ focused }: tabBarProps) => (focused
-          ? <View style={style.iconContainer}>
-            <ProfileSelectedIcon />
-            <Text style={style.iconText}>Profil</Text>
-          </View>
-          : <ProfileIcon style={style.iconContainer} />),
-      }} />
-    </Tab.Navigator>
-  );
-};
+import Home from '../Home';
 
 const MainStack = createStackNavigator();
