+++ src/Home/styles.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+import { PINK } from '../styles/colors';
+import { FIRA_SANS_BOLD } from '../styles/fonts';
+import { MARGIN, TAB_BAR_HEIGHT, TAB_BAR_LABEL_WIDTH } from '../styles/metrics';
+
+const styles = StyleSheet.create({
+  tabBar: {
+    height: TAB_BAR_HEIGHT,
+  },
+  iconContainer: {
+    alignItems: 'center',
+    width: TAB_BAR_LABEL_WIDTH,
+    marginVertical: MARGIN.SM,
+  },
+  iconText: {
+    ...FIRA_SANS_BOLD.SM,
+    color: PINK[500],
+  },
+});
+
+export default styles;
